requirements.txt
# Amenify Support Bot — Python dependencies
# Install: pip install -r requirements.txt

# ── Web framework ────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── SSE streaming ────────────────────────────────────
sse-starlette==2.1.0

# ── LLM providers ────────────────────────────────────
# Google Gemini (optional — set LLM_PROVIDER=gemini)
google-generativeai>=0.7.0
# Ollama HTTP client (optional — set LLM_PROVIDER=ollama)
httpx>=0.27.0

# ── Vector search & Cloud Embeddings ────────────────────────────────────
pinecone-client>=4.0.0
huggingface_hub>=0.20.0
transformers>=4.51.0

# ── Numerical ────────────────────────────────────────
numpy==1.26.4

# ── Web scraping (Firecrawl) ──────────────────────────
firecrawl-py>=1.5.0
# kept for any utility helpers that may reference bs4
requests==2.32.3
beautifulsoup4==4.12.3

# ── Data validation ──────────────────────────────────
pydantic==2.9.2

# ── Environment & async file I/O ─────────────────────
python-dotenv==1.0.1
aiofiles==24.1.0
